Add files via upload
Completed implementation of PGA files & requisite foundational classes/interface definitions.  Roadblock encountered due to misunderstanding of materials, and natural stopping point for the study term found there.  Multivector set theory was misunderstood to be uniform, leading to unexpected output taken as errors, whereas the algorithm for computing the vector inverse of non-versor multivector data simply turns out to be less trivial than implemented, so the PRNG input was to blame for output errrors, not the algorithm.